pyproject.toml
[project]
name = "spectralvit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "altair>=6.0.0",
    "bs4>=0.0.2",
    "graphviz>=0.21",
    "haar-pytorch>=2021.11.16.0",
    "ignite>=1.1.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "onnxsim>=0.4.36",
    "opencv-python>=4.12.0.88",
    "ort>=1.2.1",
    "packaging>=24.1",
    "pandas>=2.3.3",
    "polars>=1.36.1",
    "ptwt>=1.0.1",
    "pytest>=9.0.2",
    "pytorch-ignite>=0.5.3",
    "pytorch-warmup>=0.2.0",
    "pytorch-wavelets>=1.3.0",
    "pywavelets>=1.9.0",
    "scikit-learn>=1.8.0",
    "tensorboard>=2.20.0",
    "termcolor>=3.3.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchmetrics>=1.0.3",
    "torchview>=0.2.7",
    "torchvision>=0.24.1",
    "tqdm>=4.66.5",
    "wheel>=0.45.1",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu128"
default = false
